// typings/machines/sync/sync-writer.d.ts
export interface IBind extends IBindBase {
  (
    event: 'Enabled_enter',
    listener: () => boolean | undefined,
    context?: Object
  ): this
  (event: 'Enabled_state', listener: () => any, context?: Object): this
}
// ...
export declare type TStates =
  | 'Enabled'
  | 'Initializing'
  | 'Ready'
  | 'ConfigSet'
  | 'SubsReady'
  | 'SubsInited'
  | 'Reading'
  | 'ReadingDone'
  | 'Writing'
  | 'WritingDone'
// ...
export interface IState extends IStateBase<TStates> {}
/** Subclassable typesafe state interface */
export interface IStateExt<T extends string> extends IStateBase<T | TStates> {}
export interface IBind extends IBindBase {
  (event: TTransitions, listener: () => boolean | void, context?: Object): this
}
export interface IEmit extends IEmitBase {
  (event: TTransitions): boolean | void
}
export interface IJSONStates {
  Enabled: IState
  Initializing: IState
  Ready: IState
  ConfigSet: IState
  SubsReady: IState
  SubsInited: IState
  Reading: IState
  ReadingDone: IState
  Writing: IState
  WritingDone: IState
  Exception?: IState
}